Validate tests of Amenity entity
